Now the PATCH endpoint in Scores Controller works perfectly

+++ BackendSAP/Modelos/Calificaciones.cs
@@ -11,7 +11,7 @@ namespace BackendSAP.Modelos
         [Required]
         public string observacion { get; set; }
         [Required]
-        public ushort puntaje { get; set; }
+        public int puntaje { get; set; }
         public string usuarioId { get; set; } // Clave Foránea del usuario quién hace la calificación
         public string psicologoId { get; set; } // Clave Foránea del usuario psicólogo quién recibe la calificación
